Handle ILDA blanking metadata + opt to ignore
Sends blank points when the blank flag is set on a point, regardless of
whatever color the point encodes.

There is a command line option to override this and send whatever the
original color is specified to be. (If it is black, this option doesn't
help visualize that -- future improvement.)